Improve error handling for timeout errors
This PR improves the error handling for timeout errors. There are effectively three classes of error that the Talon client can raise:
* A request times out from Typhoeus. This generally only occurs if the client has specified a timeout.
* A request fails with an error from `libcurl`. This can occur for a host of reasons, but it is difficult to get more precise information out of `libcurl`.
* A request fails with an HTTP error. In this case, we can pull detailed response information into the error object.

With this PR, it is now possible for clients to differentiate these cases and handle them differently:
* In a timeout, `code` will be `nil` and `timeout` will be true.
* In a libcurl failure, `code` will specifically be 0.
* In an HTTP failure, the code will be something meaningful.

This isn't perfect, but it does improve the error handling ability for clients somewhat!